Fix to gamma spectrum deserialize list
